This fixes #7
Now there is the field study_server shown to the user, it is an editable field so that the user can select tables/resources from the same server and put them on different study servers to do a pooled study. Later I will add a help button to exaplain that. There is some filters implemented to check integrity of study server names